refactor: 검색 부분 타입 분리 #64
