+++ src/app/(pages)/(search)/_api/supabase.ts
@@ -1,48 +1,7 @@
+import { Club, EggClub, EggPop, PopularEggClub, WishItem } from "@/types/search.types";
 import browserClient from "@/utils/supabase/client";
 
-export type WishItem = {
-  r_c_id: {
-    regular_club_id: number; // 클럽 ID
-    regular_club_name: string; // 클럽 이름
-    regular_club_image: string; // 클럽 이미지 URL
-    regular_club_people_limited: number; // 클럽 정원
-    user_id: {
-      user_name: string; // 사용자 이름
-      user_profile_img: string; // 사용자 프로필 이미지 URL
-    }; // 사용자 배열
-    r_c_member: {
-      count: number; // 회원 수
-    }[];
-    wish_list: {
-      r_c_id: number; // 클럽 ID
-      user_id: string; // 사용자 ID
-      wish_list_id: number; // 위시리스트 ID
-    }[];
-  };
-};
-
-// PopularClub 타입 정의
-export type PopularClub = {
-  regular_club_id: number; // 클럽 ID
-  regular_club_name: string; // 클럽 이름
-  regular_club_image: string; // 클럽 이미지 URL
-  regular_club_people_limited: number; // 클럽 정원
-  user_id: {
-    user_name: string; // 사용자 이름
-    user_profile_img: string; // 사용자 프로필 이미지 URL
-  }; // 사용자 배열
-  r_c_member: {
-    count: number; // 회원 수
-  }[]; // 회원 배열
-  count: number;
-  wish_list: {
-    r_c_id: number; // 클럽 ID
-    user_id: string; // 사용자 ID
-    wish_list_id: number; // 위시리스트 ID
-  }[]; // 각 클럽의 위시리스트 개수
-};
-
-export const getPopularClubs = async (): Promise<PopularClub[]> => {
+export const getPopularClubs = async (): Promise<PopularEggClub[]> => {
   const { data, error } = await browserClient.from("wish_list").select(`
       r_c_id (
         regular_club_id,
@@ -64,7 +23,7 @@ export const getPopularClubs = async (): Promise<PopularClub[]> => {
 
   const wishes = data as unknown as WishItem[];
 
-  const grouped = wishes.reduce<Record<number, PopularClub>>((acc, curr) => {
+  const grouped = wishes.reduce<Record<number, PopularEggClub>>((acc, curr) => {
     const clubId = curr.r_c_id.regular_club_id;
     if (!acc[clubId]) {
       acc[clubId] = {
@@ -81,54 +40,6 @@ export const getPopularClubs = async (): Promise<PopularClub[]> => {
     .sort((a, b) => b.count - a.count)
     .slice(0, 10);
 };
-
-// 정규 모임과 일회성 모임의 공통 타입 정의
-interface BaseClub {
-  m_c_id: number;
-  s_c_id: number;
-  user_id: {
-    user_profile_img: string;
-    user_name: string;
-  };
-}
-
-// 정규 모임 타입 정의
-interface RegularClub extends BaseClub {
-  type: "regular";
-  regular_club_id: number;
-  regular_club_name: string;
-  regular_club_age: number;
-  regular_club_gender: string;
-  regular_club_people_limited: number;
-  regular_club_image: string;
-  regular_club_introduction: string;
-  regular_club_create_at: string;
-  regular_club_approval: boolean;
-  r_c_member: { count: number }[];
-  wish_list: {
-    r_c_id: number;
-    user_id: string;
-    wish_list_id: number;
-  }[];
-}
-
-// 일회성 모임 타입 정의
-interface OneTimeClub extends BaseClub {
-  type: "oneTime";
-  one_time_club_id: number;
-  one_time_club_name: string;
-  one_time_age: number;
-  one_time_gender: string | null;
-  one_time_people_limited: number;
-  one_time_image: string;
-  one_time_club_introduction: string;
-  one_time_create_at: string;
-  one_time_club_date_time: string;
-  one_time_club_location: string;
-  one_time_tax: number;
-}
-
-type Club = RegularClub | OneTimeClub;
 
 export const getSearchedClubs = async (searchTerm: string): Promise<Club[]> => {
   const { data: regularClubs, error: regularClubError } = await browserClient
@@ -152,12 +63,12 @@ export const getSearchedClubs = async (searchTerm: string): Promise<Club[]> => {
     throw new Error("검색 중 에러가 발생했습니다.");
   }
 
-  const typedRegularClubs: RegularClub[] = (regularClubs || []).map((club) => ({
+  const typedRegularClubs: EggClub[] = (regularClubs || []).map((club) => ({
     ...club,
     type: "regular" as const
   }));
 
-  const typedOneTimeClubs: OneTimeClub[] = (oneTimeClubs || []).map((club) => ({
+  const typedOneTimeClubs: EggPop[] = (oneTimeClubs || []).map((club) => ({
     ...club,
     type: "oneTime" as const
   }));
